+++ block-streamer/Cargo.toml
@@ -24,6 +24,8 @@ tokio-stream = "0.1.14"
 tonic = "0.10.2"
 wildmatch = "2.1.1"
 
+registry-types = { path = "../registry/types" }
+
 near-lake-framework = "0.7.4"
 
 [build-dependencies]
