+++ block-streamer/src/rules/matcher.rs
@@ -2,9 +2,9 @@ use near_lake_framework::near_indexer_primitives::{
     views::{ActionView, ExecutionStatusView, ReceiptEnumView},
     IndexerExecutionOutcomeWithReceipt,
 };
+use registry_types::{MatchingRule, Status};
 
-use crate::rules::types::events::Event;
-use crate::rules::{MatchingRule, Status};
+use crate::rules::types::Event;
 
 pub fn matches(
     matching_rule: &MatchingRule,
