+++ block-streamer/src/rules/mod.rs
@@ -3,62 +3,9 @@ pub mod outcomes_reducer;
 pub mod types;
 
 use near_lake_framework::near_indexer_primitives::StreamerMessage;
-use types::indexer_rule_match::{ChainId, IndexerRuleMatch};
+use registry_types::{IndexerRule, MatchingRule};
 
-#[cfg(not(feature = "near-sdk"))]
-use borsh::{self, BorshDeserialize, BorshSerialize};
-#[cfg(not(feature = "near-sdk"))]
-use serde::{Deserialize, Serialize};
-
-#[cfg(feature = "near-sdk")]
-use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
-#[cfg(feature = "near-sdk")]
-use near_sdk::serde::{Deserialize, Serialize};
-
-#[derive(Clone, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, PartialEq, Eq)]
-pub struct IndexerRule {
-    pub indexer_rule_kind: IndexerRuleKind,
-    pub matching_rule: MatchingRule,
-    pub id: Option<u32>,
-    pub name: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, PartialEq, Eq)]
-pub enum IndexerRuleKind {
-    Action,
-    Event,
-    AnyBlock,
-    Shard,
-}
-// future: ComposedRuleKind for multiple actions or events
-
-#[derive(Clone, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, PartialEq, Eq)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub enum Status {
-    Any,
-    Success,
-    Fail,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, PartialEq, Eq)]
-#[serde(tag = "rule", rename_all = "SCREAMING_SNAKE_CASE")]
-pub enum MatchingRule {
-    ActionAny {
-        affected_account_id: String,
-        status: Status,
-    },
-    ActionFunctionCall {
-        affected_account_id: String,
-        status: Status,
-        function: String,
-    },
-    Event {
-        contract_account_id: String,
-        standard: String,
-        version: String,
-        event: String,
-    },
-}
+use types::{ChainId, IndexerRuleMatch};
 
 pub fn reduce_indexer_rule_matches(
     indexer_rule: &IndexerRule,
