+++ indexer/indexer_rule_type/Cargo.toml
@@ -6,11 +6,5 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-borsh = { version = "0.10.2", optional = true }
-serde = { version = "1", features = ["derive"], optional = true }
-
-near-sdk = { version = "4.1.1", optional = true }
-
-[features]
-default = ["borsh", "serde"]
-nearsdk = ["near-sdk"]
+borsh = "0.10.2"
+serde = "1"
