+++ indexer/indexer_rule_type/src/indexer_rule.rs
@@ -1,12 +1,5 @@
-#[cfg(not(feature = "near-sdk"))]
 use borsh::{self, BorshDeserialize, BorshSerialize};
-#[cfg(not(feature = "near-sdk"))]
 use serde::{Deserialize, Serialize};
-
-#[cfg(feature = "near-sdk")]
-use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
-#[cfg(feature = "near-sdk")]
-use near_sdk::serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, PartialEq, Eq)]
 pub struct IndexerRule {
