+++ registry/types/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "registry-types"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+borsh = { version = "1.2.1", features = ["derive"], optional = true}
+serde = { version = "1.0.193", optional = true }
+
+near-primitives = { version = "0.17.0", optional = true}
+near-sdk = { version = "5.0.0-alpha.1", optional = true }
+
+[features]
+default = ["near-primitives", "borsh", "serde"]
+nearsdk = ["near-sdk"]
